[nix][home][linux][screen-lock] Add xautolock service to handle sleep
Without using a desktop environment or session manager, logind believes
that the session is not graphical and it is non-trivial to set the
session type ourself. See link below for more info.
https://unix.stackexchange.com/q/209274

Instead of relying on logind for sleep-on-idle, we can utilize xautolock
to do so instead. The code is copied from screen-locker.nix as there is
no way to configure both xss-lock and xautolock with home-manager
currently as hm considers them mutually exclusive.